ruby style conditions

+++ backuper.rb
@@ -9,9 +9,7 @@ backup_dir = './backups'
 file_prefix = 'hb_'
 date_format = '%Y.%m.%d'
 
-unless (File.directory?( backup_dir ))
-	File.makedirs backup_dir
-end
+File.makedirs backup_dir unless File.directory?( backup_dir )
 
 filename = "#{backup_dir}/#{file_prefix}#{Time.new.strftime( date_format )}.sql"
 included_arguments = %W[--compact --skip-triggers --no-create-info]
@@ -21,8 +19,6 @@ system command
 
 if (File.exists?( filename ))
 	system "tar czf #{filename}.tar.gz #{filename}"
-	if (File.exists?( "#{filename}.tar.gz" ))
-		File.unlink( filename )
-	end
+	File.unlink( filename ) if File.exists?( "#{filename}.tar.gz" )
 end
